Rename Common headers and update includes
