+++ AnalysisTools/DefaultAnalysis_tool.cc
@@ -12,15 +12,15 @@
 #include "nusimdata/SimulationBase/MCFlux.h"
 #include "larcore/Geometry/Geometry.h"
 #include "larsim/EventWeight/Base/MCEventWeight.h"
-#include "../Common/BacktrackingFuncs.h"
-#include "../Common/Geometry.h"
-#include "../Common/SCECorrections.h"
-#include "../Common/Containment.h"
-#include "../Common/TrackShowerScoreFuncs.h"
-#include "../Common/ProximityClustering.h"
-#include "../Common/Descendents.h"
-#include "../Common/Scatters.h"
-#include "../Common/Pandora.h"
+#include "../Common/BacktrackingUtilities.h"
+#include "../Common/GeometryUtils.h"
+#include "../Common/SpaceChargeCorrections.h"
+#include "../Common/TruthContainment.h"
+#include "../Common/TrackShowerScore.h"
+#include "../Common/HitProximityClustering.h"
+#include "../Common/PFParticleDescendants.h"
+#include "../Common/ParticleScattering.h"
+#include "../Common/PandoraUtilities.h"
 #include "ubobj/Optical/UbooneOpticalFilter.h"
 #include "canvas/Persistency/Common/TriggerResults.h"
 #include "larpandora/LArPandoraInterface/LArPandoraHelper.h"
