+++ Common/BacktrackingUtilities.h
@@ -1,5 +1,5 @@
-#ifndef BACTRACKINGFUNCS_H
-#define BACTRACKINGFUNCS_H
+#ifndef BACKTRACKINGUTILITIES_H
+#define BACKTRACKINGUTILITIES_H
 
 #include "larevt/SpaceChargeServices/SpaceChargeService.h"
 #include "lardata/DetectorInfoServices/DetectorPropertiesService.h"
