+++ Common/HitProximityClustering.h
@@ -12,8 +12,8 @@
 
     @{*/
 
-#ifndef PROXIMITYCLUSTERING_H
-#define PROXIMITYCLUSTERING_H
+#ifndef HITPROXIMITYCLUSTERING_H
+#define HITPROXIMITYCLUSTERING_H
 
 #include <map>
