+++ Common/LLRCorrectionLookup.h
@@ -1,5 +1,5 @@
-#ifndef CORRECTION_LOOKUP_H
-#define CORRECTION_LOOKUP_H
+#ifndef LLRCORRECTIONLOOKUP_H
+#define LLRCORRECTIONLOOKUP_H
 #include <stdlib.h>
 #include <vector>
