+++ Common/LLRProtonMuonLookup.h
@@ -1,5 +1,5 @@
-#ifndef PROTON_MUON_LOOKUP_H
-#define PROTON_MUON_LOOKUP_H
+#ifndef LLRPROTONMUONLOOKUP_H
+#define LLRPROTONMUONLOOKUP_H
 #include <stdlib.h>
 #include <vector>
