+++ Common/ParticleScattering.h
@@ -1,5 +1,5 @@
-#ifndef SCATTERSFUNCS_H
-#define SCATTERSFUNCS_H
+#ifndef PARTICLESCATTERING_H
+#define PARTICLESCATTERING_H
 
 #include "larpandora/LArPandoraInterface/LArPandoraHelper.h"
 #include "larpandora/LArPandoraInterface/LArPandoraGeometry.h"
