+++ Common/SpaceChargeCorrections.h
@@ -1,5 +1,5 @@
-#ifndef SCECORRECTIONSFUNCS_H
-#define SCECORRECTIONSFUNCS_H
+#ifndef SPACECHARGECORRECTIONS_H
+#define SPACECHARGECORRECTIONS_H
 
 #include "larevt/SpaceCharge/SpaceCharge.h"
 #include "larevt/SpaceChargeServices/SpaceChargeService.h"
