+++ Common/TrackShowerScore.h
@@ -1,5 +1,5 @@
-#ifndef TRACKSHOWERSCOREFUNCS_H
-#define TRACKSHOWERSCOREFUNCS_H
+#ifndef TRACKSHOWERSCORE_H
+#define TRACKSHOWERSCORE_H
 
 namespace common
 {
